netty 서버 어노테이션 주석처리 및 회원가입 기능 구현

+++ src/main/java/com/example/Dokkaebi/Repository/BatteryRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Repository
-public interface BatteryRepository extends JpaRepository<Battery,Integer>{
+public interface BatteryRepository extends JpaRepository<Battery,Long>{
 
 }
